继续文件选择，调整状态栏，修改我的页面布局等

+++ app/src/main/java/com/example/chuanke/chuanke/bean/OrderBean.java
@@ -56,7 +56,7 @@ public class OrderBean {
     private int fid;
     private String opic;
     private String ono;
-    private int osum;
+    private double osum;
 
     public void setOid(int oid) {
         this.oid = oid;
@@ -166,7 +166,7 @@ public class OrderBean {
         this.osum = osum;
     }
 
-    public int getOsum() {
+    public double getOsum() {
         return osum;
     }
